Consolidate smart capture speedtest adapter tests
